Applied Ben's changes
refs: #2302
